Rewrote order of phase shift and corrfun operations
Now, the corrfun! is applied with no offset, then, the image is FFT'ed.
Next, the gridding convolution functions are used to sample the dense
visibilities to the locations of the data set. THEN, the phase shift is
applied to these model visibilities.

Previously, the corrfun! was applied with an offset, the image was
FFT'ed, the dense visibilities were phase shifted, and then they were
sampled using the gridding convolution functions.

It turns out this seems to be all due to a misunderstanding with how I
was doing the testing. I was comparing the raw output from the FFT to
the analytical formula, when in reality I also needed to resample the
grid, too.